+++ src/components/index.js
@@ -1,23 +1,7 @@
 /**
  * Point of contact for component modules.
  * @module components
- * @example import { Field } from 'components';
  */
-
-/**
-Proposals of tree for components folder
-
-components
- |-blocks
- |-controlpanels
- |-views
- |-widgets
- |-Toolbar
- |-Header
-
-or completely flat...
-
-*/
 
 //  Do not lazy load them, since it has not much sense (they will live in the main chunk)
 export App from '@plone/volto/components/theme/App/App';